+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to Rainbow-Node-SDK will be documented in this file.
 -   Fix documentation about line feed in HTML.
 -   Replace `sendInitialBubblePresence` with `sendInitialBubblePresenceSync` method to take in to account the event received from server about the presence in the bubble.
 -   Fix missing mime.lookup with mime lib version > 1.5.0 .
+-   Fix delete / leave bubbles methods in BubblesService when the bubble has the property isActive egal to false.
 
 ## [2.7.0] - 2021-12-0
 -   Update `CallLogService` `ChannelsService` `CallLogService` `ChannelsService` `FileStorage` with categories in the documentation.